requirements.txt
flask
gunicorn
email-validator
python-dotenv
